shoper/requirements.txt
aiogram
gspread 
oauth2client
python-dotenv
sqlalchemy
aiosqlite
redis
aiohttp